[Core] Add order type in paimon (#5020)
